backend/pyproject.toml
[tool.poetry]
name = "suna"
version = "1.0"
description = "open source generalist AI Agent"
authors = ["marko-kraemer <mail@markokraemer.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://www.suna.so/"
repository = "https://github.com/kortix-ai/suna"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "1.0.1"
click = "8.1.7"
questionary = "2.0.1"
requests = "^2.31.0"
packaging = "24.1"
setuptools = "75.3.0"
pytest = "8.3.3"
pytest-asyncio = "0.24.0"
asyncio = "3.4.3"
altair = "4.2.2"
prisma = "0.15.0"
fastapi = "0.110.0"
uvicorn = "0.27.1"
python-multipart = "0.0.20"
pyjwt = "2.10.1"
exa-py = "^1.9.1"
e2b-code-interpreter = "^1.2.0"
certifi = "2024.2.2"
python-ripgrep = "0.0.6"
nest-asyncio = "^1.6.0"
vncdotool = "^1.2.0"
pytesseract = "^0.3.13"
stripe = "^12.0.1"
dramatiq = "^1.17.1"
pika = "^1.3.2"
prometheus-client = "^0.21.1"
google-generativeai = "^0.5.0"
beautifulsoup4 = "^4.12.3"
duckduckgo_search = "^5.0.0"
docker = "^7.0.0"

[tool.poetry.scripts]
agentpress = "agentpress.cli:main"

[[tool.poetry.packages]]
include = "agentpress"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
